Updated to .NET 6 and C# 10, applied some C# improvements (#1326)
Part of a series of updates to bring Catalyst and its dependencies up to date.
Updates to C# 10 and .NET 6.0. A number of Nuget packages updated to the latest, necessary since some have had security improvements.
Tested with a locally updated Nethermind code base. The Nethermind code base will be tested further and then updated to match.

CICD to be overhauled but for now these changes are necessary as part of an ongoing update programme.