Updated copyright and added additional contributor
Picking up work to update Catalyst since there are a number of projects seeking to pick the project up for NFT use cases. A DAO using DAOhaus is being formed to take over governance as a community led project along with new contributors but in the first instance the project needs to be updated, tested and put back into a condition suitable for continued community development.